requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    ./bin/compile_dependencies.sh
#
alabaster==0.7.12
    # via sphinx
amqp==5.1.1
    # via kombu
asgiref==3.5.2
    # via django
async-timeout==4.0.2
    # via redis
attrs==22.1.0
    # via jsonschema
babel==2.10.3
    # via sphinx
billiard==3.6.4.0
    # via celery
bleach==5.0.1
    # via django-markup
build==0.8.0
    # via pip-tools
celery==5.2.7
    # via -r requirements/base.in
certifi==2022.6.15
    # via
    #   django-simple-certmanager
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.0
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cryptography==38.0.1
    # via
    #   cryptograpy
    #   pyopenssl
cryptograpy==0.0.0
    # via django-simple-certmanager
deprecated==1.2.13
    # via redis
django==3.2.14
    # via
    #   -r requirements/base.in
    #   django-admin-index
    #   django-axes
    #   django-choices
    #   django-filter
    #   django-hijack
    #   django-markup
    #   django-privates
    #   django-redis
    #   django-relativedelta
    #   django-rest-framework-condition
    #   django-rosetta
    #   django-sendfile2
    #   django-simple-certmanager
    #   django-solo
    #   djangorestframework
    #   drf-nested-routers
    #   drf-spectacular
    #   notifications-api-common
    #   vng-api-common
    #   zgw-consumers
django-admin-index==2.0.2
    # via -r requirements/base.in
django-axes==5.36.0
    # via -r requirements/base.in
django-better-admin-arrayfield==1.4.2
    # via -r requirements/base.in
django-choices==1.7.2
    # via
    #   -r requirements/base.in
    #   django-simple-certmanager
    #   vng-api-common
    #   zgw-consumers
django-compat==1.0.15
    # via django-hijack-admin
django-cors-middleware==1.5.0
    # via -r requirements/base.in
django-filter==21.1
    # via
    #   -r requirements/base.in
    #   vng-api-common
django-hijack==3.2.1
    # via
    #   -r requirements/base.in
    #   django-hijack-admin
django-hijack-admin==2.1.10
    # via -r requirements/base.in
django-ipware==4.0.2
    # via django-axes
django-markup[all_filter_dependencies]==1.5
    # via -r requirements/base.in
django-ordered-model==3.6
    # via django-admin-index
django-privates==1.4.0
    # via django-simple-certmanager
django-redis==5.2.0
    # via -r requirements/base.in
django-relativedelta==1.1.2
    # via
    #   -r requirements/base.in
    #   zgw-consumers
django-rest-framework-condition==0.1.1
    # via vng-api-common
django-rosetta==0.9.8
    # via -r requirements/base.in
django-sendfile2==0.7.0
    # via django-privates
django-simple-certmanager==1.1.0
    # via zgw-consumers
django-solo==2.0.0
    # via
    #   notifications-api-common
    #   vng-api-common
    #   zgw-consumers
djangorestframework==3.12.4
    # via
    #   -r requirements/base.in
    #   drf-nested-routers
    #   drf-spectacular
    #   notifications-api-common
    #   vng-api-common
djangorestframework-camel-case==1.3.0
    # via
    #   -r requirements/base.in
    #   notifications-api-common
    #   vng-api-common
docutils==0.17.1
    # via
    #   django-markup
    #   python-creole
    #   sphinx
    #   sphinx-rtd-theme
drf-flex-fields==0.9.8
    # via -r requirements/base.in
drf-nested-routers==0.93.4
    # via vng-api-common
drf-spectacular==0.24.2
    # via
    #   -r requirements/base.in
    #   vng-api-common
drf-writable-nested==0.6.4
    # via -r requirements/base.in
faker==15.1.1
    # via zgw-consumers
gemma-zds-client==1.0.1
    # via
    #   notifications-api-common
    #   vng-api-common
    #   zgw-consumers
html5lib==1.1
    # via textile
idna==3.3
    # via requests
imagesize==1.4.1
    # via sphinx
importlib-metadata==4.12.0
    # via
    #   markdown
    #   sphinx
inflection==0.5.1
    # via drf-spectacular
iso-639==0.4.5
    # via vng-api-common
isodate==0.6.1
    # via vng-api-common
jinja2==3.1.2
    # via sphinx
jsonschema==4.16.0
    # via drf-spectacular
kombu==5.2.4
    # via celery
markdown==3.4.1
    # via
    #   -r requirements/base.in
    #   django-markup
markupsafe==2.1.1
    # via jinja2
notifications-api-common==0.1.0
    # via vng-api-common
oyaml==1.0
    # via vng-api-common
packaging==21.3
    # via
    #   build
    #   redis
    #   sphinx
pep517==0.12.0
    # via build
pika==1.3.0
    # via -r requirements/base.in
pip-tools==6.8.0
    # via -r requirements/base.in
polib==1.1.1
    # via django-rosetta
prompt-toolkit==3.0.30
    # via click-repl
psycopg2==2.9.3
    # via -r requirements/base.in
pycparser==2.21
    # via cffi
pygments==2.12.0
    # via
    #   django-markup
    #   sphinx
pyjwt==2.4.0
    # via
    #   gemma-zds-client
    #   vng-api-common
pyopenssl==22.1.0
    # via
    #   django-simple-certmanager
    #   zgw-consumers
pyparsing==3.0.9
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
python-creole==1.4.10
    # via django-markup
python-dateutil==2.8.2
    # via
    #   -r requirements/base.in
    #   django-relativedelta
    #   faker
python-dotenv==0.20.0
    # via -r requirements/base.in
pytz==2022.1
    # via
    #   -r requirements/base.in
    #   babel
    #   celery
    #   django
pyyaml==6.0
    # via
    #   drf-spectacular
    #   gemma-zds-client
    #   oyaml
    #   vng-api-common
raven==6.10.0
    # via -r requirements/base.in
redis==4.3.4
    # via django-redis
regex==2022.7.25
    # via textile
requests==2.28.1
    # via
    #   django-rosetta
    #   gemma-zds-client
    #   requests-mock
    #   sphinx
    #   vng-api-common
    #   zgw-consumers
requests-mock==1.10.0
    # via zgw-consumers
six==1.16.0
    # via
    #   bleach
    #   click-repl
    #   django-choices
    #   django-compat
    #   html5lib
    #   isodate
    #   python-dateutil
    #   requests-mock
smartypants==2.0.1
    # via django-markup
snowballstemmer==2.2.0
    # via sphinx
sphinx==5.0.2
    # via
    #   -r requirements/base.in
    #   sphinx-rtd-theme
sphinx-rtd-theme==1.0.0
    # via -r requirements/base.in
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sqlparse==0.4.2
    # via django
textile==4.0.2
    # via django-markup
tomli==2.0.1
    # via
    #   build
    #   pep517
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.10
    # via requests
uwsgi==2.0.20
    # via -r requirements/base.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
vng_api_common @ git+https://github.com/VNG-Realisatie/vng-api-common@drf-spectacular-upgrade
    # via -r requirements/base.in
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   html5lib
wheel==0.37.1
    # via pip-tools
wrapt==1.14.1
    # via deprecated
zgw-consumers==0.21.2
    # via notifications-api-common
zipp==3.8.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
pip==22.1.2
    # via pip-tools
setuptools==63.2.0
    # via
    #   django-axes
    #   pip-tools
